[追加]DIspose()処理を共通化
