+++ Niconicome/ViewModels/BindableBase.cs
@@ -3,11 +3,17 @@ using System.Collections.Generic;
 using System.Text;
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
+using System.Reactive.Disposables;
 
 namespace Niconicome.ViewModels
 {
-    public class BindableBase : INotifyPropertyChanged
+    public class BindableBase : INotifyPropertyChanged, IDisposable
     {
+        ~BindableBase()
+        {
+            this.Dispose();
+        }
+
         public event PropertyChangedEventHandler? PropertyChanged;
 
         protected virtual void OnPropertyChanged([CallerMemberName] string? name = null)
@@ -21,6 +27,18 @@ namespace Niconicome.ViewModels
             field = value;
             this.OnPropertyChanged(name);
             return true;
+        }
+
+        protected CompositeDisposable disposables = new();
+
+        protected bool hasDisposed;
+
+        public void Dispose()
+        {
+            if (this.hasDisposed) return;
+            this.disposables.Dispose();
+            this.hasDisposed = true;
+            GC.SuppressFinalize(this);
         }
     }
